+++ tests/Predis/Connection/ParametersTest.php
@@ -30,7 +30,6 @@ class ParametersTest extends PredisTestCase
         $this->assertEquals($defaults['scheme'], $parameters->scheme);
         $this->assertEquals($defaults['host'], $parameters->host);
         $this->assertEquals($defaults['port'], $parameters->port);
-        $this->assertEquals($defaults['timeout'], $parameters->timeout);
     }
 
     /**
@@ -285,7 +284,6 @@ class ParametersTest extends PredisTestCase
             'scheme' => 'tcp',
             'host' => '127.0.0.1',
             'port' => 6379,
-            'timeout' => 5.0,
         );
     }